requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
sentence-transformers==3.0.1
pgvector==0.2.5